requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
cryptography==2.4.2
emoji==0.5.1
ephem==3.7.6.0
future==0.17.1
idna==2.7
pycparser==2.19
PySocks==1.6.8
python-telegram-bot==11.1.0
six==1.11.0